Extract presentation primitive drawing
